+++ Source/RhythmCodex.Lib/Iso/Streamers/ICdSectorStreamReader.cs
@@ -6,6 +6,6 @@ namespace RhythmCodex.Iso.Streamers
 {
     public interface ICdSectorStreamReader
     {
-        IEnumerable<ICdSector> Read(Stream stream, int length, bool keepOnDisk);
+        IEnumerable<ICdSector> Read(Stream stream, long length, bool keepOnDisk);
     }
 }
